+++ app/gui/integration-test/project-view/editorPanels.spec.ts
@@ -9,6 +9,7 @@ import * as locate from './locate'
 async function goToGraphAndGetDocs(editorPage: EditorPageActions) {
   let docsContent: Locator
   let docsScroller: Locator
+  await editorPage.toggleDocsAssetPanel()
   await editorPage.do(async (page) => {
     docsContent = page.getByTestId('documentation-editor-content')
     docsScroller = page.getByTestId('documentation-editor-scroller')
@@ -44,14 +45,18 @@ test.describe('Main method documentation rendering', () => {
         .locator('span.cm-BulletList-item span')
         .getByText(text, { exact: true })
         .boundingBox()
-    const listLevel0 = await listItemPos('Outer list element')
-    const listLevel1 = await listItemPos('Nested list element')
-    const listLevel2 = await listItemPos('Very nested list element')
+    const listLevel0 = () => listItemPos('Outer list element')
+    const listLevel1 = () => listItemPos('Nested list element')
+    const listLevel2 = () => listItemPos('Very nested list element')
     expect(listLevel0).not.toBeNull()
     expect(listLevel1).not.toBeNull()
     expect(listLevel2).not.toBeNull()
-    expect(listLevel0!.x).toBeLessThan(listLevel1!.x)
-    expect(listLevel1!.x).toBeLessThan(listLevel2!.x)
+    expect
+      .poll(async () => ((await listLevel1())?.x ?? NaN) - ((await listLevel0())?.x ?? NaN))
+      .toBeGreaterThan(0)
+    expect
+      .poll(async () => ((await listLevel2())?.x ?? NaN) - ((await listLevel1())?.x ?? NaN))
+      .toBeGreaterThan(0)
   })
 
   test('Link (rendered and interactive)', async ({ editorPage, page, context }) => {